feat: modules to mjs
